requirements.txt
langchain
langchain_core
langchain_community
langchain-openai
langchain-text-splitters
unstructured[md] # markdown loading
chromadb 
tiktoken 
dspy-ai[chromadb]  
python-dotenv 
ipykernel
tqdm